+++ resources/views/sshkey/_spalten.blade.php
@@ -1,0 +1,3 @@
+{{-- Spaltenueberschriften dieser Tabelle. Der Schluessel selbst steht nicht
+     hier - er ist mehrere Zeilen lang und gehoert ins Bearbeiten-Fenster. --}}
+<x-table.head :labels="['Name', 'Benutzername', 'Verfahren', 'Beschreibung', 'Verwendet bei', '', ]" />