+++ resources/views/sshkey/index.blade.php
@@ -1,0 +1,5 @@
+<x-app-layout :$customer>
+    {{-- Liste und Formular sind Livewire: Ein im Modal angelegter Eintrag
+         erscheint sofort, ohne die Seite neu zu laden. --}}
+    <livewire:objekt-liste typ="sshkey" :customer="$customer" />
+</x-app-layout>